requirements.txt
flask
scikit-learn
joblib
numpy
requests
python-dotenv
